moved limiters to their own classes, deleted higher_order_funcs.py
